Sample file and better organization
